requirements.txt
alembic==1.6.0
attrs==21.2.0
certifi==2021.10.8
chardet==4.0.0
click==8.0.3
coverage==5.5
Faker==8.1.2
factory-boy==3.2.0
Flask==1.1.2
Flask-Injector==0.12.3
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==2.10
iniconfig==1.1.1
injector==0.18.4
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
packaging==21.3
pluggy==0.13.1
PyMySQL==1.0.2
py==1.11.0
pydantic==1.8.1
pyparsing==3.0.6
pytest==6.2.4
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==0.17.1
python-editor==1.0.4
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.13
SQLAlchemy-Utils==0.37.8
text-unidecode==1.3
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==1.0.1
